+++ Saurons/Heistotron/src/main/java/com/sauron/heist/heistron/orchestration/Taskium.java
@@ -7,5 +7,5 @@ public interface Taskium extends Processum {
         return (ChildHeistOrchestrator) this.getTaskManager();
     }
 
-    int getTaskId();
+    long getTaskId();
 }
